ODROID-N2: Wifi: enable to load multiple driver module per a device
/vendor/etc/wifi_id_list.txt stores WiFi device's VID & PID as a reference to
judge the device driver modules, it indicates the driver list to load as per
detected dongle. Each line consists with 4 parts, 'VID'/'PID'/'NICK'/'PATH',
The file in 'PATH' will be read to load drivers if matched VID and PID are
found.

The file 'PATH' lists the driver modules to load in order of its dependency
and unloaded its bardward.

Change-Id: I1ae2b3607c28c01dddf7ea356d41b8c358a19639